PM-29806: Move FlightRecorderWriter to the data module
