+++ data/src/main/kotlin/com/bitwarden/data/manager/di/DataManagerModule.kt
@@ -1,6 +1,7 @@
 package com.bitwarden.data.manager.di
 
 import android.content.Context
+import com.bitwarden.core.data.manager.BuildInfoManager
 import com.bitwarden.core.data.manager.dispatcher.DispatcherManager
 import com.bitwarden.data.manager.BitwardenPackageManager
 import com.bitwarden.data.manager.BitwardenPackageManagerImpl
@@ -8,12 +9,15 @@ import com.bitwarden.data.manager.NativeLibraryManager
 import com.bitwarden.data.manager.NativeLibraryManagerImpl
 import com.bitwarden.data.manager.file.FileManager
 import com.bitwarden.data.manager.file.FileManagerImpl
+import com.bitwarden.data.manager.flightrecorder.FlightRecorderWriter
+import com.bitwarden.data.manager.flightrecorder.FlightRecorderWriterImpl
 import com.bitwarden.network.service.DownloadService
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
+import java.time.Clock
 import javax.inject.Singleton
 
 /**
@@ -39,6 +43,20 @@ object DataManagerModule {
         context = context,
         downloadService = downloadService,
         dispatcherManager = dispatcherManager,
+    )
+
+    @Provides
+    @Singleton
+    fun provideFlightRecorderWriter(
+        clock: Clock,
+        fileManager: FileManager,
+        dispatcherManager: DispatcherManager,
+        buildInfoManager: BuildInfoManager,
+    ): FlightRecorderWriter = FlightRecorderWriterImpl(
+        clock = clock,
+        fileManager = fileManager,
+        dispatcherManager = dispatcherManager,
+        buildInfoManager = buildInfoManager,
     )
 
     @Provides
